requirements.txt
wikipedia==1.4.0
requests==2.24.0
translate==3.5.0
discord.py==1.6.0
discord==1.0.1
python-dotenv==0.17.1
